+++ packages/native/src/components/containers/Box/index.tsx
@@ -11,7 +11,7 @@ export const Box = makeComponent(
 )
 
 export const pullBoxProps = makePropPuller(boxConfig)
-export const getBoxStyle = makeStyleGetter(boxConfig)
+export const getBoxStyles = makeStyleGetter(boxConfig)
 
 export * from './types'
 export * from './config'