+++ packages/native/src/components/content/Text/index.tsx
@@ -7,7 +7,7 @@ import { textConfig } from './config'
 export const Text = makeComponent(RNText, textConfig)
 
 export const pullTextProps = makePropPuller(textConfig)
-export const getTextStyle = makeStyleGetter(textConfig)
+export const getTextStyles = makeStyleGetter(textConfig)
 
 export * from './types'
 export * from './config'